Pull out lib_no_cgo.go
